requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pdf2image==1.17.0
pillow==11.2.1
pytesseract==0.3.13
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.1.3
